requirements.txt
Flask==2.2.2
PyGithub==1.55
redis==4.1.0
gunicorn==20.1.0
celery==5.2.7
